requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
requests==2.33.0
gunicorn==22.0.0
whitenoise==6.8.0
